+++ NEWS.md
@@ -1,3 +1,10 @@
+# kgp 1.1.0
+
+- Added new `allmeta` tibble with population and region metadata for populations in the 1000 Genomes Project (kgp), Simons Genome Diversity Project (sgdp), Human Genome Diversity Project (hgdp), and Gambian Genome Variation Project (ggvp)
+- Added latitude and longitude information to the 1000 Genomes Project metadata (kgpmeta)
+- Made region names from the 1000 Genomes Project consistent with region names from other data collections (e.g., "European" to "Europe," "East Asian" to "East Asia," etc.)
+- Updated documentation
+
 # kgp 1.0.0
 
 - Initial release